Add files via upload
i was able to get the j48 and random forest to work with cross validation and sampling with smote.  i was able to do this by removing some of the derived variables.  there appears to be something in the derived variables causing problems.  i still cannot get the hour parsing to work from the click_time column